+++ 2019/Day05.cs
@@ -12,7 +12,7 @@ namespace adventofcode
             List<string> input = File.ReadAllLines(Path.Combine(Environment.CurrentDirectory, "input", "day05.txt")).ToList();
             foreach(string intcode in input)
             {
-                int[] opcodes = intcode.Split(",").ToList().Select(o => Int32.Parse(o)).ToArray();
+                long[] opcodes = intcode.Split(",").ToList().Select(o => long.Parse(o)).ToArray();
                 IntCode intCode = new IntCode(opcodes);
                 Console.WriteLine(intCode.Compute(5).Dequeue());
             }
